requirements.txt
# to install as pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu121
deepinv==0.3.0
diffusers==0.32.2
huggingface_hub==0.25.2
hydra-core==1.3.2
numpy==1.26.4
omegaconf==2.3.0
Pillow==10.2.0
safetensors==0.5.3
scipy==1.11.4
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.48.0
open-clip-torch==2.23.0
accelerate==1.8.1
peft>=0.7.0
pandas==2.3.0